requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
pytz>=2021.3
requests>=2.25.1
pymongo>=4.0.0